buenoshabitos

+++ practica1/src/PD.c
@@ -2,25 +2,18 @@
 #include <stdlib.h>
 #include <string.h>
 #include <unistd.h>
-#include <signal.h>
+#include <dirent.h>
 
-void manejador(int sig);
 void removeDirectory(void){
 
-    system("rm -rf ./utils/estudiantes");
+    //Comprobamos que el directorio estudiantes existe y lo borramos
+    if(opendir("./utils/estudiantes")){
+        system("rm -rf ./utils/estudiantes");
+    }  
 } 
 
 int main (){
-    FILE *file;
-    char DNI [8];
-    signal(SIGINT, &manejador);
     removeDirectory();
     exit(0);
     return 0;
-}
-
-
-void manejador(int sig){
-    printf("Process kill: %d\n",getpid());
-    exit(0);
 }
